The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Provinces Outbreak Dataset

daily summary of the Coronavirus (COVID-19) pandemic cases in Italy's provinces

italy_province

Format

A data.frame object

Source

Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile website

Details

The dataset contains the daily summary of the Coronavirus pandemic cases in Italy's regions. The data includes the following fields:

- date - date of notification (Date class)

- region_name - region names

- region_code - region code

- province_name - province names

- province_spatial - the spatial province names as in the output of the `ne_states` function from the **rnaturalearth** package

- province_abb - the province abbreviation

- province_code - province code

- lat - latitude point

- long - Longitude point

- new_cases - daily number of positive cases

- total_cases - total number of positive cases (cumulative)

- nuts_code_1 - groups of region - level 1 of Italy statistical regions (Nomenclature of Territorial Units for Statistics), more details available here

- nuts_code_2 - regions, level 2 of Italy statistical regions (Nomenclature of Territorial Units for Statistics), more details available here

- nuts_code_3 - provinces, level 3 of Italy statistical regions (Nomenclature of Territorial Units for Statistics), more details available here

Examples

data(italy_province)

head(italy_province)

#> # A tibble: 6 x 14
#>   date       region_name region_code province_name province_spatial province_abb
#>   <date>     <chr>       <chr>       <chr>         <chr>            <chr>       
#> 1 2020-02-24 Abruzzo     13          L'Aquila      L'Aquila         "AQ"        
#> 2 2020-02-24 Abruzzo     13          Teramo        Teramo           "TE"        
#> 3 2020-02-24 Abruzzo     13          Pescara       Pescara          "PE"        
#> 4 2020-02-24 Abruzzo     13          Chieti        Chieti           "CH"        
#> 5 2020-02-24 Abruzzo     13          In fase di d… In fase di defi… ""          
#> 6 2020-02-24 Basilicata  17          Potenza       Potenza          "PZ"        
#> # … with 8 more variables: province_code <chr>, lat <dbl>, long <dbl>,
#> #   new_cases <dbl>, total_cases <dbl>, nuts_code_1 <chr>, nuts_code_2 <chr>,
#> #   nuts_code_3 <chr>