The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Regions Outbreak Dataset

daily summary of the Coronavirus (COVID-19) pandemic cases in Italy's regions

italy_region

Format

A data.frame object

Source

Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile website

Details

The dataset contains the daily summary of the Coronavirus pandemic cases in Italy's regions. The data includes the following fields:

- date - date of notification (Date class)

- region_code - region code

- region_name - region names

- lat - latitude point

- long - Longitude point

- hospitalized_with_symptoms - number of hospitalised patients with symptoms

- intensive_care - intensive Care

- total_hospitalized - total hospitalised patients

- home_confinement - home confinement

- total_currently_positive - total amount of current positive cases (Hospitalised patients + Home confinement)

- new_currently_positive - news amount of current positive cases (Hospitalised patients + Home confinement)

- recovered - recovered cases

- death - death cases

- total_positive_cases - total positive cases

- total_tests - number of tests performed

- region_spatial - the spatial region name as in the output of the `ne_states` function from the **rnaturalearth** packag

Examples

data(italy_region)

head(italy_region)

#>         date region_code    region_name      lat     long
#> 1 2020-02-24          13        Abruzzo 42.35122 13.39844
#> 2 2020-02-24          17     Basilicata 40.63947 15.80515
#> 3 2020-02-24           4   P.A. Bolzano 46.49933 11.35662
#> 4 2020-02-24          18       Calabria 38.90598 16.59440
#> 5 2020-02-24          15       Campania 40.83957 14.25085
#> 6 2020-02-24           8 Emilia Romagna 44.49437 11.34172
#>   hospitalized_with_symptoms intensive_care total_hospitalized home_confinement
#> 1                          0              0                  0                0
#> 2                          0              0                  0                0
#> 3                          0              0                  0                0
#> 4                          0              0                  0                0
#> 5                          0              0                  0                0
#> 6                         10              2                 12                6
#>   total_currently_positive new_currently_positive recovered death
#> 1                        0                      0         0     0
#> 2                        0                      0         0     0
#> 3                        0                      0         0     0
#> 4                        0                      0         0     0
#> 5                        0                      0         0     0
#> 6                       18                     18         0     0
#>   total_positive_cases total_tests      region_spatial
#> 1                    0           5             Abruzzo
#> 2                    0           0          Basilicata
#> 3                    0           1 Trentino-Alto Adige
#> 4                    0           1            Calabria
#> 5                    0          10            Campania
#> 6                   18         148      Emilia-Romagna