The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Outbreak Dataset

daily summary of the Coronavirus (COVID-19) pandemic cases in Italy

italy_total

Format

A data.frame object

Source

Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile website

Details

The dataset contains the daily summary of the Coronavirus pandemic cases in Italy. The data includes the following fields:

- date - date of notification (Date class)

- hospitalized_with_symptoms - number of hospitalised patients with symptoms

- intensive_care - intensive Care

- total_hospitalized - total hospitalised patients

- home_confinement - home confinement

- cumulative_positive_cases - total amount of current positive cases (Hospitalised patients + Home confinement)

- daily_positive_cases - new amount of current positive cases (Hospitalised patients + Home confinement)

- recovered - recovered cases

- death - death cases

- positive_clinical_activity - positive cases emerged from clinical activity

- positive_surveys_tests - positive cases emerging from surveys and tests, planned at national or regional level

- cumulative_cases - total positive cases, recovered, and death cases

- total_tests - number of tests performed

- total_people_tested - total number of people tested

Examples

data(italy_total)

head(italy_total)

#>         date hospitalized_with_symptoms intensive_care total_hospitalized
#> 1 2020-02-24                        101             26                127
#> 2 2020-02-25                        114             35                150
#> 3 2020-02-26                        128             36                164
#> 4 2020-02-27                        248             56                304
#> 5 2020-02-28                        345             64                409
#> 6 2020-02-29                        401            105                506
#>   home_confinement cumulative_positive_cases daily_positive_cases recovered
#> 1               94                       221                    0         1
#> 2              162                       311                   90         1
#> 3              221                       385                   74         3
#> 4              284                       588                  203        45
#> 5              412                       821                  233        46
#> 6              543                      1049                  228        50
#>   death positive_clinical_activity positive_surveys_tests cumulative_cases
#> 1     7                         NA                     NA              229
#> 2    10                         NA                     NA              322
#> 3    12                         NA                     NA              400
#> 4    17                         NA                     NA              650
#> 5    21                         NA                     NA              888
#> 6    29                         NA                     NA             1128
#>   total_tests total_people_tested
#> 1        4324                  NA
#> 2        8623                  NA
#> 3        9587                  NA
#> 4       12014                  NA
#> 5       15695                  NA
#> 6       18661                  NA