The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Provinces Outbreak Dataset

daily summary of the Coronavirus (COVID-19) pandemic cases in Italy's provinces

italy_province

Format

A data.frame object

Source

Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile website

Details

The dataset contains the daily summary of the Coronavirus pandemic cases in Italy's regions. The data includes the following fields:

- date - date of notification (Date class)

- region_code - region code

- region_name - region names

- province_code - province code

- province_name - province names

- province_abb - the province abbreviation

- lat - latitude point

- long - Longitude point

- total_cases - total number of positive cases (cumulative)

- new_cases - daily number of positive cases

- province_spatial - the spatial province names as in the output of the `ne_states` function from the **rnaturalearth** package

Examples

data(italy_province)

head(italy_province)

#>         date region_code region_name province_code
#> 1 2020-02-24          13     Abruzzo            66
#> 2 2020-02-24          13     Abruzzo            67
#> 3 2020-02-24          13     Abruzzo            68
#> 4 2020-02-24          13     Abruzzo            69
#> 5 2020-02-24          13     Abruzzo           979
#> 6 2020-02-24          17  Basilicata            76
#>                          province_name province_abb      lat     long
#> 1                             L'Aquila           AQ 42.35122 13.39844
#> 2                               Teramo           TE 42.65892 13.70440
#> 3                              Pescara           PE 42.46458 14.21365
#> 4                               Chieti           CH 42.35103 14.16755
#> 5 In fase di definizione/aggiornamento                    NA       NA
#> 6                              Potenza           PZ 40.63947 15.80515
#>   total_cases new_cases                     province_spatial
#> 1           0         0                             L'Aquila
#> 2           0         0                               Teramo
#> 3           0         0                              Pescara
#> 4           0         0                               Chieti
#> 5           0         0 In fase di definizione/aggiornamento
#> 6           0         0                              Potenza