The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Regions Outbreak Dataset

daily summary of the Coronavirus (COVID-19) pandemic cases in Italy's regions

italy_region

Format

A data.frame object

Source

Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile website

Details

The dataset contains the daily summary of the Coronavirus pandemic cases in Italy's regions. The data includes the following fields:

- date - date of notification (Date class)

- region_code - region code

- region_name - region names

- lat - latitude point

- long - Longitude point

- hospitalized_with_symptoms - number of hospitalised patients with symptoms

- intensive_care - intensive Care

- total_hospitalized - total hospitalised patients

- home_confinement - home confinement

- cumulative_positive_cases - total amount of current positive cases (Hospitalised patients + Home confinement)

- daily_positive_cases - news amount of current positive cases (Hospitalised patients + Home confinement)

- recovered - recovered cases

- death - death cases

- positive_clinical_activity - positive cases emerged from clinical activity

- positive_surveys_tests - positive cases emerging from surveys and tests, planned at national or regional level

- cumulative_cases - total positive, recovered, and death cases

- total_tests - number of tests performed

- total_people_tested - total number of people tested

- region_spatial - the spatial region names as in the output of the `ne_states` function from the **rnaturalearth** package

Examples

data(italy_region)

head(italy_region)

#>         date region_code           region_name      lat     long
#> 1 2020-02-24          13               Abruzzo 42.35122 13.39844
#> 2 2020-02-24          17            Basilicata 40.63947 15.80515
#> 3 2020-02-24          18              Calabria 38.90598 16.59440
#> 4 2020-02-24          15              Campania 40.83957 14.25085
#> 5 2020-02-24           8        Emilia-Romagna 44.49437 11.34172
#> 6 2020-02-24           6 Friuli Venezia Giulia 45.64944 13.76814
#>   hospitalized_with_symptoms intensive_care total_hospitalized home_confinement
#> 1                          0              0                  0                0
#> 2                          0              0                  0                0
#> 3                          0              0                  0                0
#> 4                          0              0                  0                0
#> 5                         10              2                 12                6
#> 6                          0              0                  0                0
#>   cumulative_positive_cases daily_positive_cases recovered death
#> 1                         0                    0         0     0
#> 2                         0                    0         0     0
#> 3                         0                    0         0     0
#> 4                         0                    0         0     0
#> 5                        18                    0         0     0
#> 6                         0                    0         0     0
#>   positive_clinical_activity positive_surveys_tests cumulative_cases
#> 1                         NA                     NA                0
#> 2                         NA                     NA                0
#> 3                         NA                     NA                0
#> 4                         NA                     NA                0
#> 5                         NA                     NA               18
#> 6                         NA                     NA                0
#>   total_tests total_people_tested        region_spatial
#> 1           5                  NA               Abruzzo
#> 2           0                  NA            Basilicata
#> 3           1                  NA              Calabria
#> 4          10                  NA              Campania
#> 5         148                  NA        Emilia-Romagna
#> 6          58                  NA Friuli-Venezia Giulia